Update after some modifications
Hard- and Firmware of Controller_X0.1  now work together. The first 8 Valves are fine, two on the other side are problematic.  The software now works together with the Firmware of Controller_X. Unfortunately, The Software seems to freeze after a while, when Picarro data is read.

Trouble shooting is ongoing.

Documentation follows as soon as the setup works reliably.